Add files via upload
The program takes a block of text from the user and converts it to an array. It then asks the user to pick a letter and generates a poem using that letter by iterating over the provided block of text to find all the words that begin with the users chosen letter.